added formatTime method

+++ src/main/java/com/jodexindustries/freecases/utils/CooldownManager.java
@@ -5,9 +5,9 @@ import java.util.Map;
 import java.util.UUID;
 
 public class CooldownManager {
-    public static final Map<UUID, Integer> cooldowns = new HashMap<>();
+    public static final Map<UUID, Long> cooldowns = new HashMap<>();
 
-    public static void setCooldown(UUID player, int time){
+    public static void setCooldown(UUID player, long time){
         if(time < 1) {
             cooldowns.remove(player);
         } else {
@@ -15,7 +15,7 @@ public class CooldownManager {
         }
     }
 
-    public static int getCooldown(UUID player){
-        return cooldowns.getOrDefault(player, 0);
+    public static long getCooldown(UUID player){
+        return cooldowns.getOrDefault(player, 0L);
     }
 }
